Moved ArcGisTokenGenerator and ArcGisTokenManager to maplayers-auth packages,
