+++ core/frontend/src/tile/internal.ts
@@ -51,8 +51,6 @@ export * from "./B3dmReader";
 export * from "./ImdlReader";
 export * from "./map/ArcGISTileMap";
 export * from "./map/MapLayerAuthentication";
-export * from "./map/ArcGisTokenGenerator";
-export * from "./map/ArcGisTokenManager";
 export * from "./map/MapFeatureInfo";
 export * from "./map/MapLayerFormatRegistry";
 export * from "./map/ArcGisUtilities";
