+++ extensions/map-layers-auth/src/map-layers-auth.ts
@@ -6,3 +6,5 @@ export * from "./ArcGis/ArcGisOAuth2Endpoint";
 export * from "./ArcGis/ArcGisUrl";
 export * from "./ArcGis/ArcGisAccessClient";
 export * from "./ArcGis/ArcGisOauthRedirect";
+export * from "./ArcGis/ArcGisTokenGenerator";
+export * from "./ArcGis/ArcGisTokenManager";
